+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased] - TBD
 ### Changed
 - Moved function `make_benchmark_conc_plots` from `gcpy/benchmark/modules/benchmark_funcs.py` to `gcpy/benchmark/modules/benchmark_conc.py`
+- Moved function `make_benchmark_emis_plots` from `gcpy/benchmark/modules/benchmark_funcs.py` to `gcpy/benchmark/modules/benchmark_emissions.py`
+- Moved function `make_benchmark_emis_tables` from `gcpy/benchmark/modules/benchmark_funcs.py` to `gcpy/benchmark/modules/benchmark_emissions.py`
+- Moved function `create_total_emissions_tables` from `gcpy/benchmark/modules/benchmark_funcs.py` to `gcpy/benchmark/modules/benchmark_emissions.py`
 
 ### Fixed
 - Passed `spcdb_files` as the 7th argument (following `dev_interval`) to `make_benchmark_operations_budget`, where missing
@@ -21,7 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 - Fixed error in `get_input_res` for numpy 2.x versions; we first copy the lon and lat arrays before sorting
-
+b
 ## [1.7.0] - 2026-01-29
 ### Added
 - Added `gcpy/profile/vtune_plot_hotspots.py` to plot a bargraph of hotspots from Intel VTune reports